style: clippy 1.94 対応（新規 lint の警告一掃）
- cargo clippy --fix による機械修正（io::Error::other、collapsible_if、
  manual_strip、map_or 簡約、needless borrow、is_multiple_of、derivable_impls 等）
- 手修正: uninit_vec（l4 転送バッファ、理由付き allow）、manual c-string、
  lines().flatten()→map_while、same_item_push→resize、match→matches!、
  strip_prefix、len+is_empty、from_str→parse_str 改名（FromStr 混同回避）、
  single_match→if let、&PathBuf→&Path、フィールド代入→構造体初期化、
  const アサーション化、迷子の cfg(grpc-web) 属性除去（e2e_tests）
- 意図的パターンへの理由付き allow: drop_non_drop（借用解放の明示）、
  vec_box（SendMsgState のアドレス固定）、large_enum_variant（HostPortStr の
  スタック整形）、await_holding_refcell_ref（H3 メインループの単一借用者設計）、
  await_holding_lock（リロード E2E の直列化ガード）

+++ src/wasm/host/grpc.rs
@@ -292,7 +292,7 @@ fn proxy_grpc_call_impl(
     // Write call_id to return pointer
     if return_call_id_ptr > 0 {
         if let Some(memory) = caller.get_export("memory").and_then(|e| e.into_memory()) {
-            let call_id_bytes = (call_id as u32).to_le_bytes();
+            let call_id_bytes = call_id.to_le_bytes();
             let data = memory.data_mut(&mut *caller);
             let ptr = return_call_id_ptr as usize;
             if ptr + 4 <= data.len() {
